Added wcs_fliplist
This wordlist/dictionary maker, can only be used for Wilson County Schools email passwords. Wilson County has a common password vulnerability. Teachers and students default passwords consists of the following, Doe0000 , Doe22131. These are characterized by first three initials of their last name, and followed by 5 digits as their lunch codes. if paired with hydra, this wordlist can be a powerful tool for exploiting their known and vulnerable accounts. This wordlist asks for their First three initials of their last name which the first letter has to be in caps, and adds numbers to the end starting from 0 to 99999. Of course I have to include the "DO NOT USE THIS FOR NEFARIOUS PURPOSES", but we all know what happens, happens. Good luck!